bugfix(modify-node hook): the third argument was not supplied
I also recognized the semantices of the third argument make no sense
and I fixed them.

Previously I suppled ``node.call`` as 3rd argument. But ``node`` is
provided anyways. What might be useful instead is the name of the filter.
Because if you reuse the same Lua function for multiple hooks, you might
want to know which filter this function was called for.